feat(eleicao): change form names
