--- app/eleicao/forms/__init__.py
@@ -1,10 +1,0 @@
-from .create import (
-    Candidate1Form,
-    Candidate2Form,
-    Candidate3Form,
-    Candidate4Form,
-    Candidate5Form,
-    Candidate6Form,
-    Candidate7Form,
-    VoterForm
-)
